pyproject.toml
[tool.poetry]
name = "victoria_metrics_jobs"
version = "0.1.3"
description = "Victoria Metrics Jobs - A job scheduler service for Victoria Metrics"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "victoria_metrics_jobs"}]
readme = "README.md"
homepage = "https://github.com/your-org/victoria_metrics_jobs"
repository = "https://github.com/your-org/victoria_metrics_jobs"
keywords = ["victoria-metrics", "scheduler", "monitoring", "jobs"]

[tool.poetry.dependencies]
python = "^3.9"
apscheduler = "^3.10.4"
cron-descriptor = "^1.2.30"
omegaconf = "^2.3.0"
requests = "^2.31.0"
sqlalchemy = "^2.0.0"
protobuf = "^5.29.0"
python-snappy = "^0.7.3"
prometheus-api-client = "0.6.0"
prometheus-client = "^0.20.0"
flask = "^3.0.0"
gunicorn = "^22.0.0"
waitress = "^3.0.0"
pandas = "^2.1.0"
prophet = "^1.1.5"
papermill = "^2.6.0"
darts = "^0.39.0"
scipy = "^1.11.0"
nbconvert = "^7.16.0"
pygments = "^2.17.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"

[tool.poetry.scripts]
victoria_metrics_jobs = "main:main"

# Deployment and service files
[tool.victoria_metrics_jobs]
service_file = "scripts/victoria_metrics_jobs.service"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Test configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "--cov=scheduler",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests"
]

# Coverage configuration
[tool.coverage.run]
source = ["scheduler"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod"
]
